[EZ] Replace `pytorch-labs` with `meta-pytorch` (#98)
This PR replaces all instances of `pytorch-labs` with `meta-pytorch` in
this repository now that the `pytorch-labs` org has been renamed to
`meta-pytorch`

## Changes Made
- Replaced all occurrences of `pytorch-labs` with `meta-pytorch`
- Skipped binary files and files larger than 1MB due to GitHub api
payload limits in the script to cover all repos in this org. Will do a
more manual second pass later to cover any larger files

## Files Modified
This PR updates files that contained the target text.

Generated by automated script on 2025-08-22T23:26:44.714975+00:00Z